requirements.txt
asgiref==3.11.1
autopep8==2.3.2
coverage==7.13.5
Django==5.2.12
django-filter==25.2
djangorestframework==3.17.1
flake8==7.3.0
mccabe==0.7.0
pillow==12.2.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.2
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.15.0
